fix most merge bugs. Moved login files to account
