+++ com/DansFrappes/src/DjangoServer/account/templates/account/login.html
@@ -6,11 +6,11 @@
 	<head>
 		<title>TODO: change this title with JavaScript</title>
 		<meta charset="utf-8">
-		<link href="{% static 'login/style.css' %}" rel="stylesheet">
+		<link href="{% static 'account/style.css' %}" rel="stylesheet">
 	</head>
 
 	<body>
-		<script src="{% static 'login/header.js' %}"></script>
+		<script src="{% static 'account/header.js' %}"></script>
 		<form method="post" action="">{% csrf_token %}
 			<input type="text" id="username" name="username">
 			<input type="password" id="password" name="password">
@@ -21,7 +21,7 @@
 		{% else %}
 		<p id="invalidLoginText">Not Hidden</p>
 		{% endif %}
-		<script src="{% static 'login/script.js' %}"></script>
+		<script src="{% static 'account/script.js' %}"></script>
 		
 	</body>
 </html>
