+++ com/DansFrappes/src/DjangoServer/account/urls.py
@@ -3,9 +3,9 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
-    path('edit', views.edit, name='edit'),
-    path('view', views.view, name='view'),
-    path('create', views.new_user, name='create'),
-    path('login', views.user_login, name='login'),
-    path('logout', views.user_logout, name='logout')
+    path('edit/', views.edit, name='edit'),
+    path('view/', views.view, name='view'),
+    path('createaccount/', views.createaccount, name='create'),
+    path('login/', views.login, name='login'),
+    path('logout/', views.user_logout, name='logout')
 ]
